requirements.txt
aiohappyeyeballs>=2.4.0
aiohttp>=3.8
aiosignal==1.4.0
alpaca-py==0.11.0
appdirs==1.4.4
async-timeout==5.0.1
attrs>=23.1
beautifulsoup4>=4.12
blinker==1.9.0
cachetools==5.5.2
certifi>=2024.7.4
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cssselect==1.3.0
cssutils==2.11.1
curl_cffi==0.12.0
deprecation==2.1.0
Flask==3.1.1
frozendict==2.4.6
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
html5lib==1.1
idna>=3.4
importlib_metadata>=6.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml>=4.9
MarkupSafe==3.0.2
more-itertools>=10.1
msgpack==1.0.3
multidict==6.6.3
multitasking==0.0.12
numpy==2.0.2
oauthlib==3.3.1
packaging>=23.2
pandas>=2.2
peewee==3.18.2
platformdirs==4.3.8
premailer==3.10.0
propcache==0.3.2
protobuf==6.31.1
psutil>=5.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pytest>=8.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz>=2024.1
PyYAML==6.0.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
secure-smtplib==0.1.1
six>=1.16
soupsieve==2.7
typing_extensions>=4.12
tzdata>=2024.1
urllib3>=1.26,<3
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
yagmail==0.15.293
yarl>=1.9
zipp==3.23.0
websockets==11.0.3
oauth2client>=4.1.3
